# openapi_output.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Smith 2.0",
        "description": "Agente de Vendas Inteligente com IA",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/leads/": {
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Create Lead",
                "description": "Cria um novo lead\n\nArgs:\n    lead_data: Dados do lead a criar\n\nReturns:\n    Lead criado com ID",
                "operationId": "create_lead_api_leads__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LeadCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeadResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "List Leads",
                "description": "Lista leads com filtros opcionais\n\nArgs:\n    status: Filtrar por status\n    origem: Filtrar por origem\n    temperatura: Filtrar por temperatura\n    limit: N\u00c3\u00bamero m\u00c3\u00a1ximo de resultados\n    offset: Offset para pagina\u00c3\u00a7\u00c3\u00a3o\n\nReturns:\n    Lista de leads",
                "operationId": "list_leads_api_leads__get",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/LeadStatus"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    },
                    {
                        "name": "origem",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/LeadOrigin"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Origem"
                        }
                    },
                    {
                        "name": "temperatura",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/LeadTemperature"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Temperatura"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Lead"
                                    },
                                    "title": "Response List Leads Api Leads  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/leads/{lead_id}": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Get Lead",
                "description": "Busca um lead por ID\n\nArgs:\n    lead_id: ID do lead\n\nReturns:\n    Lead encontrado",
                "operationId": "get_lead_api_leads__lead_id__get",
                "parameters": [
                    {
                        "name": "lead_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Lead Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Leads"
                ],
                "summary": "Update Lead",
                "description": "Atualiza um lead existente\n\nArgs:\n    lead_id: ID do lead\n    lead_data: Dados a atualizar\n\nReturns:\n    Lead atualizado",
                "operationId": "update_lead_api_leads__lead_id__put",
                "parameters": [
                    {
                        "name": "lead_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Lead Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LeadUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeadResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Leads"
                ],
                "summary": "Delete Lead",
                "description": "Deleta um lead\n\nArgs:\n    lead_id: ID do lead\n\nReturns:\n    Confirma\u00c3\u00a7\u00c3\u00a3o de dele\u00c3\u00a7\u00c3\u00a3o",
                "operationId": "delete_lead_api_leads__lead_id__delete",
                "parameters": [
                    {
                        "name": "lead_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Lead Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeadResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/leads/{lead_id}/qualify": {
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Force Qualification",
                "description": "For\u00c3\u00a7a re-qualifica\u00c3\u00a7\u00c3\u00a3o de um lead\n\nArgs:\n    lead_id: ID do lead\n\nReturns:\n    Lead com novo score",
                "operationId": "force_qualification_api_leads__lead_id__qualify_post",
                "parameters": [
                    {
                        "name": "lead_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Lead Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeadResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LeadCreate": {
                "properties": {
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "telefone": {
                        "type": "string",
                        "title": "Telefone"
                    },
                    "empresa": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Empresa"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "origem": {
                        "$ref": "#/components/schemas/LeadOrigin"
                    },
                    "notas": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notas"
                    }
                },
                "type": "object",
                "required": [
                    "nome",
                    "telefone",
                    "origem"
                ],
                "title": "LeadCreate",
                "description": "Payload para criar um lead"
            },
            "LeadOrigin": {
                "type": "string",
                "enum": [
                    "instagram",
                    "whatsapp",
                    "site",
                    "indicacao",
                    "facebook_ads",
                    "teste",
                    "outro"
                ],
                "title": "LeadOrigin",
                "description": "Origem do lead"
            },
            "LeadResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "lead": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Lead"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "success"
                ],
                "title": "LeadResponse",
                "description": "Response padr\u00c3\u00a3o de Lead"
            },
            "LeadStatus": {
                "type": "string",
                "enum": [
                    "novo",
                    "contato_inicial",
                    "qualificando",
                    "qualificado",
                    "agendamento_marcado",
                    "ganho",
                    "perdido"
                ],
                "title": "LeadStatus",
                "description": "Status do lead no funil de vendas"
            },
            "LeadTemperature": {
                "type": "string",
                "enum": [
                    "quente",
                    "morno",
                    "frio"
                ],
                "title": "LeadTemperature",
                "description": "Temperatura do lead (engajamento)"
            },
            "LeadUpdate": {
                "properties": {
                    "nome": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Nome"
                    },
                    "empresa": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Empresa"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "status": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/LeadStatus"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "temperatura": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/LeadTemperature"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "qualification_data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/QualificationData"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "roi_analysis": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ROIAnalysis"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "valor_estimado": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Valor Estimado"
                    },
                    "notas": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notas"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags"
                    },
                    "ai_summary": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ai Summary"
                    },
                    "requires_human_approval": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Requires Human Approval"
                    }
                },
                "type": "object",
                "title": "LeadUpdate",
                "description": "Payload para atualizar um lead"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}